requirements.txt
asgiref==3.8.1
Django==5.1.5
django-tenants==3.7.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
asgiref==3.8.1
Django==5.1.5
django-tenants==3.7.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.5
django-crispy-forms==2.3
django-recaptcha==4.0.0
django-tenants==3.7.0
pillow==11.1.0
psycopg2-binary==2.9.10
python-http-client==3.3.7
pytz==2025.1
sendgrid==6.11.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.5
django-crispy-forms==2.3
django-recaptcha==4.0.0
django-tenants==3.7.0
dotenv==0.9.9
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.1
sendgrid==6.11.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
